+++ cmd/anti_affinity_group_delete.go
@@ -5,9 +5,8 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/exoscale/cli/pkg/account"
 	"github.com/exoscale/cli/pkg/globalstate"
-	exoapi "github.com/exoscale/egoscale/v2/api"
+	v3 "github.com/exoscale/egoscale/v3"
 )
 
 type antiAffinityGroupDeleteCmd struct {
@@ -33,11 +32,14 @@ func (c *antiAffinityGroupDeleteCmd) cmdPreRun(cmd *cobra.Command, args []string
 }
 
 func (c *antiAffinityGroupDeleteCmd) cmdRun(_ *cobra.Command, _ []string) error {
-	zone := account.CurrentAccount.DefaultZone
+	ctx := gContext
 
-	ctx := exoapi.WithEndpoint(gContext, exoapi.NewReqEndpoint(account.CurrentAccount.Environment, zone))
+	antiAffinityGroupsResp, err := globalstate.EgoscaleV3Client.ListAntiAffinityGroups(ctx)
+	if err != nil {
+		return err
+	}
 
-	antiAffinityGroup, err := globalstate.EgoscaleClient.FindAntiAffinityGroup(ctx, zone, c.AntiAffinityGroup)
+	antiAffinityGroup, err := antiAffinityGroupsResp.FindAntiAffinityGroup(c.AntiAffinityGroup)
 	if err != nil {
 		return err
 	}
@@ -51,14 +53,19 @@ func (c *antiAffinityGroupDeleteCmd) cmdRun(_ *cobra.Command, _ []string) error 
 		}
 	}
 
-	decorateAsyncOperation(fmt.Sprintf("Deleting Anti-Affinity Group %s...", c.AntiAffinityGroup), func() {
-		err = globalstate.EgoscaleClient.DeleteAntiAffinityGroup(ctx, zone, antiAffinityGroup)
-	})
-	if err != nil {
-		return err
-	}
+	return decorateAsyncOperations(fmt.Sprintf("Deleting Anti-Affinity Group %s...", c.AntiAffinityGroup), func() error {
+		op, err := globalstate.EgoscaleV3Client.DeleteAntiAffinityGroup(ctx, antiAffinityGroup.ID)
+		if err != nil {
+			return fmt.Errorf("exoscale: error while deleting Anti Affinity Group: %w", err)
+		}
 
-	return nil
+		_, err = globalstate.EgoscaleV3Client.Wait(ctx, op, v3.OperationStateSuccess)
+		if err != nil {
+			return fmt.Errorf("exoscale: error while waiting for Anti Affinity Group deletion: %w", err)
+		}
+
+		return nil
+	})
 }
 
 func init() {