+++ cmd/instance_update.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"strings"
 
-	egoscale3 "github.com/exoscale/egoscale/v3"
-
 	"github.com/spf13/cobra"
 
 	"github.com/exoscale/cli/pkg/account"
@@ -14,6 +12,7 @@ import (
 	"github.com/exoscale/cli/pkg/output"
 	"github.com/exoscale/cli/pkg/userdata"
 	exoapi "github.com/exoscale/egoscale/v2/api"
+	v3 "github.com/exoscale/egoscale/v3"
 )
 
 type instanceUpdateCmd struct {
@@ -102,25 +101,26 @@ func (c *instanceUpdateCmd) cmdRun(cmd *cobra.Command, _ []string) error {
 			}
 
 			if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.Protection)) {
-				var value egoscale3.UUID
-				var op *egoscale3.Operation
-				value, err = egoscale3.ParseUUID(*instance.ID)
+				var client *v3.Client
+				client, err = switchClientZoneV3(ctx, globalstate.EgoscaleV3Client, v3.ZoneName(c.Zone))
+
+				var instanceID v3.UUID
+				var op *v3.Operation
+				instanceID, err = v3.ParseUUID(*instance.ID)
 				if err != nil {
 					return
 				}
 				if c.Protection {
-					op, err = globalstate.EgoscaleV3Client.AddInstanceProtection(ctx, value)
+					op, err = client.AddInstanceProtection(ctx, instanceID)
 				} else {
-					op, err = globalstate.EgoscaleV3Client.RemoveInstanceProtection(ctx, value)
+					op, err = client.RemoveInstanceProtection(ctx, instanceID)
 				}
 				if err != nil {
 					return
 				}
-				_, err = globalstate.EgoscaleV3Client.Wait(ctx, op, egoscale3.OperationStateSuccess)
+				_, err = client.Wait(ctx, op, v3.OperationStateSuccess)
 			}
-
 		})
-
 		if err != nil {
 			return err
 		}