+++ cmd/private_network_show.go
@@ -2,20 +2,18 @@ package cmd
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
+	"net"
 	"os"
 	"strings"
 
 	"github.com/olekukonko/tablewriter"
 	"github.com/spf13/cobra"
 
-	"github.com/exoscale/cli/pkg/account"
 	"github.com/exoscale/cli/pkg/globalstate"
 	"github.com/exoscale/cli/pkg/output"
 	"github.com/exoscale/cli/table"
-	"github.com/exoscale/cli/utils"
-	exoapi "github.com/exoscale/egoscale/v2/api"
+	v3 "github.com/exoscale/egoscale/v3"
 )
 
 type privateNetworkLeaseOutput struct {
@@ -23,16 +21,24 @@ type privateNetworkLeaseOutput struct {
 	IPAddress string `json:"ip_address"`
 }
 
+type privateNetworkOptionsOutput struct {
+	Routers      []net.IP `json:"routers"`
+	DNSServers   []net.IP `json:"dns-servers"`
+	NTPServers   []net.IP `json:"ntp-servers"`
+	DomainSearch []string `json:"domain-search"`
+}
+
 type privateNetworkShowOutput struct {
-	ID          string                      `json:"id"`
+	ID          v3.UUID                     `json:"id"`
 	Name        string                      `json:"name"`
 	Description string                      `json:"description"`
-	Zone        string                      `json:"zone"`
+	Zone        v3.ZoneName                 `json:"zone"`
 	Type        string                      `json:"type"`
 	StartIP     *string                     `json:"start_ip,omitempty"`
 	EndIP       *string                     `json:"end_ip,omitempty"`
 	Netmask     *string                     `json:"netmask,omitempty"`
 	Leases      []privateNetworkLeaseOutput `json:"leases,omitempty"`
+	Options     privateNetworkOptionsOutput `json:"options"`
 }
 
 func (o *privateNetworkShowOutput) ToJSON() { output.JSON(o) }
@@ -42,35 +48,24 @@ func (o *privateNetworkShowOutput) ToTable() {
 	t.SetHeader([]string{"Private Network"})
 	defer t.Render()
 
-	t.Append([]string{"ID", o.ID})
+	t.Append([]string{"ID", o.ID.String()})
 	t.Append([]string{"Name", o.Name})
 	t.Append([]string{"Description", o.Description})
-	t.Append([]string{"Zone", o.Zone})
+	t.Append([]string{"Zone", string(o.Zone)})
 	t.Append([]string{"Type", o.Type})
 
 	if o.Type == "managed" {
 		t.Append([]string{"Start IP", *o.StartIP})
 		t.Append([]string{"End IP", *o.EndIP})
 		t.Append([]string{"Netmask", *o.Netmask})
+
 		t.Append([]string{
-			"Leases", func(leases []privateNetworkLeaseOutput) string {
-				if len(leases) > 0 {
-					buf := bytes.NewBuffer(nil)
-					at := table.NewEmbeddedTable(buf)
-					at.SetHeader([]string{" "})
-					at.SetAlignment(tablewriter.ALIGN_LEFT)
-
-					for _, lease := range leases {
-						at.Append([]string{lease.Instance, lease.IPAddress})
-					}
-					at.Render()
-
-					return buf.String()
-				}
-				return "-"
-			}(o.Leases),
+			"Leases", formatLeases(o.Leases),
 		})
 	}
+	t.Append([]string{
+		"Options", formatOptions(o.Options),
+	})
 }
 
 type privateNetworkShowCmd struct {
@@ -80,7 +75,7 @@ type privateNetworkShowCmd struct {
 
 	PrivateNetwork string `cli-arg:"#" cli-usage:"NAME|ID"`
 
-	Zone string `cli-short:"z" cli-usage:"Private Network zone"`
+	Zone v3.ZoneName `cli-short:"z" cli-usage:"Private Network zone"`
 }
 
 func (c *privateNetworkShowCmd) cmdAliases() []string { return gShowAlias }
@@ -105,22 +100,44 @@ func (c *privateNetworkShowCmd) cmdPreRun(cmd *cobra.Command, args []string) err
 }
 
 func (c *privateNetworkShowCmd) cmdRun(_ *cobra.Command, _ []string) error {
-	ctx := exoapi.WithEndpoint(gContext, exoapi.NewReqEndpoint(account.CurrentAccount.Environment, c.Zone))
-
-	privateNetwork, err := globalstate.EgoscaleClient.FindPrivateNetwork(ctx, c.Zone, c.PrivateNetwork)
+	ctx := gContext
+	client, err := switchClientZoneV3(ctx, globalstate.EgoscaleV3Client, c.Zone)
 	if err != nil {
-		if errors.Is(err, exoapi.ErrNotFound) {
-			return fmt.Errorf("resource not found in zone %q", c.Zone)
-		}
+		return err
+	}
+
+	resp, err := client.ListPrivateNetworks(ctx)
+	if err != nil {
+		return err
+	}
+
+	pn, err := resp.FindPrivateNetwork(c.PrivateNetwork)
+	if err != nil {
+		return err
+	}
+
+	privateNetwork, err := client.GetPrivateNetwork(ctx, pn.ID)
+	if err != nil {
 		return err
 	}
 
 	out := privateNetworkShowOutput{
-		ID:          *privateNetwork.ID,
+		ID:          privateNetwork.ID,
 		Zone:        c.Zone,
-		Name:        *privateNetwork.Name,
-		Description: utils.DefaultString(privateNetwork.Description, ""),
+		Name:        privateNetwork.Name,
+		Description: privateNetwork.Description,
 		Type:        "manual",
+		Options: func() privateNetworkOptionsOutput {
+			if privateNetwork.Options == nil {
+				return privateNetworkOptionsOutput{}
+			}
+			return privateNetworkOptionsOutput{
+				Routers:      privateNetwork.Options.Routers,
+				DNSServers:   privateNetwork.Options.DNSServers,
+				NTPServers:   privateNetwork.Options.NtpServers,
+				DomainSearch: privateNetwork.Options.DomainSearch,
+			}
+		}(),
 	}
 
 	if privateNetwork.StartIP != nil {
@@ -140,14 +157,14 @@ func (c *privateNetworkShowCmd) cmdRun(_ *cobra.Command, _ []string) error {
 		out.Leases = make([]privateNetworkLeaseOutput, 0)
 
 		for _, lease := range privateNetwork.Leases {
-			instance, err := globalstate.EgoscaleClient.GetInstance(ctx, c.Zone, *lease.InstanceID)
+			instance, err := client.GetInstance(ctx, lease.InstanceID)
 			if err != nil {
-				return fmt.Errorf("unable to retrieve Compute instance %s: %w", *lease.InstanceID, err)
+				return fmt.Errorf("unable to retrieve Compute instance %s: %w", lease.InstanceID, err)
 			}
 
 			out.Leases = append(out.Leases, privateNetworkLeaseOutput{
-				Instance:  *instance.Name,
-				IPAddress: lease.IPAddress.String(),
+				Instance:  instance.Name,
+				IPAddress: lease.IP.String(),
 			})
 		}
 	}
@@ -159,4 +176,60 @@ func init() {
 	cobra.CheckErr(registerCLICommand(privateNetworkCmd, &privateNetworkShowCmd{
 		cliCommandSettings: defaultCLICmdSettings(),
 	}))
+}
+
+func ipSliceToStringSlice(ips []net.IP) []string {
+	result := make([]string, len(ips))
+	for i, ip := range ips {
+		result[i] = ip.String()
+	}
+	return result
+}
+
+func formatLeases(leases []privateNetworkLeaseOutput) string {
+	if len(leases) == 0 {
+		return "-"
+	}
+
+	buf := bytes.NewBuffer(nil)
+	at := table.NewEmbeddedTable(buf)
+	at.SetHeader([]string{" "})
+	at.SetAlignment(tablewriter.ALIGN_LEFT)
+
+	for _, lease := range leases {
+		at.Append([]string{lease.Instance, lease.IPAddress})
+	}
+	at.Render()
+
+	return buf.String()
+}
+
+func formatOptions(opts privateNetworkOptionsOutput) string {
+	hasOptions := len(opts.Routers) > 0 || len(opts.DNSServers) > 0 ||
+		len(opts.NTPServers) > 0 || len(opts.DomainSearch) > 0
+
+	if !hasOptions {
+		return "-"
+	}
+
+	buf := bytes.NewBuffer(nil)
+	at := table.NewEmbeddedTable(buf)
+	at.SetHeader([]string{" "})
+	at.SetAlignment(tablewriter.ALIGN_LEFT)
+
+	if len(opts.Routers) > 0 {
+		at.Append([]string{"Routers", strings.Join(ipSliceToStringSlice(opts.Routers), ", ")})
+	}
+	if len(opts.DNSServers) > 0 {
+		at.Append([]string{"DNS Servers", strings.Join(ipSliceToStringSlice(opts.DNSServers), ", ")})
+	}
+	if len(opts.NTPServers) > 0 {
+		at.Append([]string{"NTP Servers", strings.Join(ipSliceToStringSlice(opts.NTPServers), ", ")})
+	}
+	if len(opts.DomainSearch) > 0 {
+		at.Append([]string{"Domain Search", strings.Join(opts.DomainSearch, ", ")})
+	}
+
+	at.Render()
+	return buf.String()
 }